# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API de Renta de Bicicletas
  description: Adaptador HTTP de referencia (hexagonal) del caso de uso **crear renta
    multi-bicicleta** (UC-01). Los adaptadores de salida son en memoria; el dominio
    no depende de este framework.
  version: 0.1.0
paths:
  /rentals:
    post:
      summary: Crear renta de varias bicicletas
      description: Crea una renta atómica de una o varias bicicletas de una estación
        (UC-01). Si una bicicleta no está disponible o el pago es rechazado, no se
        renta ninguna ni se cobra (RN-05).
      operationId: create_rental_endpoint_rentals_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRentalRequest'
        required: true
      responses:
        '201':
          description: Renta creada y activa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalResponse'
        '402':
          description: Pago rechazado por la pasarela
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Bicicleta o estación inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'Conflicto: bicicleta no disponible, ya rentada o tarifa inactiva'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Entrada inválida o regla de negocio (lista vacía / bicicletas
            duplicadas)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rentals/{rental_id}/returns:
    post:
      summary: Devolver una o varias bicicletas de una renta
      description: 'Devuelve una o varias bicicletas de una renta en una estación
        destino (UC-02). Devolución total → renta ''completada''; parcial → ''parcialmente_devuelta''
        (estado derivado). Operación atómica: si algo falla, no se devuelve ninguna
        (RN-05). No liquida el pago ni cobra reubicación (fuera de alcance E-04).'
      operationId: return_bicycles_endpoint_rentals__rental_id__returns_post
      parameters:
      - name: rental_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rental Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReturnBicyclesRequest'
      responses:
        '200':
          description: Renta actualizada tras la devolución
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalView'
        '404':
          description: Renta o estación destino inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'Conflicto: renta no activa, ítem ya devuelto, bicicleta ajena
            a la renta o estación destino llena'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Entrada inválida (lista vacía / UUID malformado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /stations:
    get:
      summary: Listar estaciones
      description: 'HU-10: lista todas las estaciones sembradas para descubrir ids
        e inventario. Solo lectura; lista vacía (200, []) si no hay estaciones.'
      operationId: list_stations_endpoint_stations_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/StationView'
                type: array
                title: Response List Stations Endpoint Stations Get
  /stations/{station_id}/bicycles:
    get:
      summary: Bicicletas de una estación
      description: 'HU-11: lista las bicicletas ubicadas en una estación. Con ``?available=true``
        filtra solo las disponibles en estación.'
      operationId: list_station_bicycles_endpoint_stations__station_id__bicycles_get
      parameters:
      - name: station_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Station Id
      - name: available
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Available
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BicycleView'
                title: Response List Station Bicycles Endpoint Stations  Station Id  Bicycles
                  Get
        '404':
          description: Estación inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: station_id no es un UUID válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /rentals/{rental_id}:
    get:
      summary: Consultar una renta por id
      description: 'HU-12: obtiene una renta por id para verificar su estado e ítems
        tras crearla. Solo lectura.'
      operationId: get_rental_endpoint_rentals__rental_id__get
      parameters:
      - name: rental_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rental Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RentalView'
        '404':
          description: Renta inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: rental_id no es un UUID válido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /health:
    get:
      summary: Estado del servicio
      description: Verifica que el servicio responde (sin tocar el dominio).
      operationId: health_endpoint_health_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
components:
  schemas:
    BicycleView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: Identificador de la bicicleta
        code:
          type: string
          title: Code
          description: Código de la bicicleta, p. ej. 'BIKE-1'
        status:
          type: string
          title: Status
          description: Estado de la bicicleta, p. ej. 'disponible'
      type: object
      required:
      - id
      - code
      - status
      title: BicycleView
      description: Vista de una bicicleta (HU-11).
    CreateRentalRequest:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
          description: Identificador del cliente que renta
        station_id:
          type: string
          format: uuid
          title: Station Id
          description: Estación de origen de la renta
        bicycle_ids:
          items:
            type: string
            format: uuid
          type: array
          minItems: 1
          title: Bicycle Ids
          description: Bicicletas a rentar (al menos una)
      additionalProperties: false
      type: object
      required:
      - user_id
      - station_id
      - bicycle_ids
      title: CreateRentalRequest
      description: 'Cuerpo de ``POST /rentals`` (HU-05/HU-07).


        Pydantic valida la sintaxis de los UUID (malformado → 422) y exige una lista

        de bicicletas no vacía en el borde (``min_length=1``); el dominio además lanza

        ``EmptyRentalError`` como red de seguridad. ``extra=''forbid''`` rechaza campos

        desconocidos en la frontera.'
    ErrorResponse:
      properties:
        error:
          type: string
          title: Error
          description: Nombre del error de dominio
        detail:
          type: string
          title: Detail
          description: Mensaje legible del error
      type: object
      required:
      - error
      - detail
      title: ErrorResponse
      description: 'Cuerpo uniforme de error. ``error`` es el nombre estable del error
        de

        dominio; ``detail`` es un mensaje legible — nunca una traza de error.'
    HealthResponse:
      properties:
        status:
          type: string
          title: Status
          description: Estado del servicio, p. ej. 'ok'
      type: object
      required:
      - status
      title: HealthResponse
      description: 'Cuerpo de ``GET /health`` (HU-08): estado del servicio.'
    MoneyView:
      properties:
        amount:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Amount
          description: Monto con 2 decimales (DECIMAL(12,2))
        currency:
          type: string
          title: Currency
          description: Moneda, p. ej. 'COP'
      type: object
      required:
      - amount
      - currency
      title: MoneyView
      description: 'Vista de un valor ``Money`` del dominio.


        ``amount`` es ``Decimal`` (DECIMAL(12,2)); Pydantic lo serializa como número

        JSON, NUNCA como ``float``, preservando la precisión monetaria.'
    RentalItemView:
      properties:
        bicycle_id:
          type: string
          format: uuid
          title: Bicycle Id
          description: Bicicleta asociada al ítem
        status:
          type: string
          title: Status
          description: Estado del ítem, p. ej. 'activo' o 'devuelto'
        estimated_amount:
          $ref: '#/components/schemas/MoneyView'
          description: Monto estimado del ítem
        final_amount:
          anyOf:
          - $ref: '#/components/schemas/MoneyView'
          - type: 'null'
          description: Monto final al devolver (None si aún activo)
        usage_minutes:
          anyOf:
          - type: integer
          - type: 'null'
          title: Usage Minutes
          description: Minutos de uso al devolver (None si aún activo)
        returned_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Returned At
          description: Instante de devolución (None si aún activo)
      type: object
      required:
      - bicycle_id
      - status
      - estimated_amount
      title: RentalItemView
      description: 'Vista de un ítem de renta (HU-12).


        Los campos de devolución (E-04) son nulos mientras el ítem está ``activo``
        y

        se rellenan al devolverlo (``devuelto``).'
    RentalResponse:
      properties:
        rental_id:
          type: string
          format: uuid
          title: Rental Id
          description: Identificador de la renta creada
        payment_id:
          type: string
          format: uuid
          title: Payment Id
          description: Identificador del pago autorizado
        status:
          type: string
          title: Status
          description: Estado de la renta (p. ej. 'activa')
      type: object
      required:
      - rental_id
      - payment_id
      - status
      title: RentalResponse
      description: 'Respuesta de éxito de ``POST /rentals`` (201): la renta creada.'
    RentalView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: Identificador de la renta
        status:
          type: string
          title: Status
          description: Estado de la renta, p. ej. 'activa'
        estimated_total:
          $ref: '#/components/schemas/MoneyView'
          description: Total estimado de la renta
        payment_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Payment Id
          description: Pago asociado (None si aún no confirmada / fallida)
        items:
          items:
            $ref: '#/components/schemas/RentalItemView'
          type: array
          title: Items
          description: Ítems de la renta
      type: object
      required:
      - id
      - status
      - estimated_total
      - items
      title: RentalView
      description: Vista de una renta y sus ítems (HU-12).
    ReturnBicyclesRequest:
      properties:
        bicycle_ids:
          items:
            type: string
            format: uuid
          type: array
          minItems: 1
          title: Bicycle Ids
          description: Bicicletas a devolver (al menos una)
        return_station_id:
          type: string
          format: uuid
          title: Return Station Id
          description: Estación destino de la devolución
      additionalProperties: false
      type: object
      required:
      - bicycle_ids
      - return_station_id
      title: ReturnBicyclesRequest
      description: 'Cuerpo de ``POST /rentals/{rental_id}/returns`` (HU-16/UC-02).


        Pydantic valida la sintaxis de los UUID (malformado → 422) y exige una lista

        de bicicletas no vacía (``min_length=1``); el dominio además lanza

        ``EmptyRentalError``/``DuplicateBicycleError`` como red de seguridad.

        ``extra=''forbid''`` rechaza campos desconocidos en la frontera. El

        ``rental_id`` viaja como parámetro de ruta, no en el cuerpo.'
    StationView:
      properties:
        id:
          type: string
          format: uuid
          title: Id
          description: Identificador de la estación
        code:
          type: string
          title: Code
          description: Código de la estación, p. ej. 'ST-01'
        name:
          type: string
          title: Name
          description: Nombre de la estación
        capacity:
          type: integer
          title: Capacity
          description: Capacidad total de la estación
        available_inventory:
          type: integer
          title: Available Inventory
          description: Bicicletas disponibles en la estación
      type: object
      required:
      - id
      - code
      - name
      - capacity
      - available_inventory
      title: StationView
      description: Vista de una estación (HU-10).